# [core] Speed up dask tests using a shared Ray instance fixture