--- a/python/ray/util/dask/scheduler.py
+++ b/python/ray/util/dask/scheduler.py
@@ -1,18 +1,19 @@
 import atexit
 import threading
+import time
 from collections import defaultdict
 from collections import OrderedDict
 from dataclasses import dataclass
 from multiprocessing.pool import ThreadPool
+from pprint import pprint
 from typing import Optional
-
-import ray
 
 import dask
 from dask.core import istask, ishashable, _execute_task
 from dask.system import CPU_COUNT
 from dask.threaded import pack_exception, _thread_get_id
 
+import ray
 from ray.util.dask.callbacks import local_ray_callbacks, unpack_ray_callbacks
 from ray.util.dask.common import unpack_object_refs
 from ray.util.dask.scheduler_utils import get_async, apply_sync
@@ -459,14 +460,11 @@ def render_progress_bar(tracker, object_refs):
         )
         if len(ready_refs) == len(object_refs):
             break
-        import time
-
         time.sleep(0.1)
     pb_bar.close()
     submitted, finished = ray.get(tracker.result.remote())
     if submitted != finished:
         print("Completed. There was state inconsistency.")
-    from pprint import pprint
 
     pprint(ray.get(tracker.report.remote()))