--- a/python/ray/util/dask/tests/test_dask_callback.py
+++ b/python/ray/util/dask/tests/test_dask_callback.py
@@ -1,22 +1,17 @@
 import sys
 
+import sys
+
 import dask
 import pytest
 
 import ray
+from ray.tests.conftest import *  # noqa: F403, F401
 from ray.util.dask import ray_dask_get, RayDaskCallback
 
 pytestmark = pytest.mark.skipif(
     sys.version_info >= (3, 12), reason="Skip dask tests for Python version 3.12+"
 )
-
-
-@pytest.fixture
-def ray_start_1_cpu():
-    address_info = ray.init(num_cpus=2)
-    yield address_info
-    # The code after the yield will run as teardown code.
-    ray.shutdown()
 
 
 @dask.delayed
@@ -34,7 +29,7 @@ def test_callback_active():
     assert not RayDaskCallback.ray_active
 
 
-def test_presubmit_shortcircuit(ray_start_1_cpu):
+def test_presubmit_shortcircuit(ray_start_regular_shared):
     """
     Test that presubmit return short-circuits task submission, and that task's
     result is set to the presubmit return value.
@@ -57,7 +52,7 @@ def test_presubmit_shortcircuit(ray_start_1_cpu):
     assert result == 0
 
 
-def test_pretask_posttask_shared_state(ray_start_1_cpu):
+def test_pretask_posttask_shared_state(ray_start_regular_shared):
     """
     Test that pretask return value is passed to corresponding posttask
     callback.
@@ -77,7 +72,7 @@ def test_pretask_posttask_shared_state(ray_start_1_cpu):
     assert result == 5
 
 
-def test_postsubmit(ray_start_1_cpu):
+def test_postsubmit(ray_start_regular_shared):
     """
     Test that postsubmit is called after each task.
     """
@@ -109,7 +104,7 @@ def test_postsubmit(ray_start_1_cpu):
     assert result == 5
 
 
-def test_postsubmit_all(ray_start_1_cpu):
+def test_postsubmit_all(ray_start_regular_shared):
     """
     Test that postsubmit_all is called once.
     """
@@ -141,7 +136,7 @@ def test_postsubmit_all(ray_start_1_cpu):
     assert result == 5
 
 
-def test_finish(ray_start_1_cpu):
+def test_finish(ray_start_regular_shared):
     """
     Test that finish callback is called once.
     """
@@ -173,7 +168,7 @@ def test_finish(ray_start_1_cpu):
     assert result == 5
 
 
-def test_multiple_callbacks(ray_start_1_cpu):
+def test_multiple_callbacks(ray_start_regular_shared):
     """
     Test that multiple callbacks are supported.
     """
@@ -208,7 +203,7 @@ def test_multiple_callbacks(ray_start_1_cpu):
     assert result == 5
 
 
-def test_pretask_posttask_shared_state_multi(ray_start_1_cpu):
+def test_pretask_posttask_shared_state_multi(ray_start_regular_shared):
     """
     Test that pretask return values are passed to the correct corresponding
     posttask callbacks when multiple callbacks are given.
@@ -244,6 +239,4 @@ def test_pretask_posttask_shared_state_multi(ray_start_1_cpu):
 
 
 if __name__ == "__main__":
-    import sys
-
     sys.exit(pytest.main(["-v", __file__]))