--- a/python/ray/util/dask/tests/test_dask_callback.py
+++ b/python/ray/util/dask/tests/test_dask_callback.py
@@ -1,6 +1,5 @@
 import sys
 
-import sys
 
 import dask
 import pytest