--- a/python/ray/util/dask/tests/test_dask_multi_node.py
+++ b/python/ray/util/dask/tests/test_dask_multi_node.py
@@ -1,17 +1,10 @@
 import sys
 
 import dask
-import dask.array as da
-import dask.dataframe as dd
-import numpy as np
-import pandas as pd
 import pytest
 
 import ray
 from ray.tests.conftest import *  # noqa: F403, F401
-from ray.util.client.common import ClientObjectRef
-from ray.util.dask import disable_dask_on_ray, enable_dask_on_ray, ray_dask_get
-from ray.util.dask.callbacks import ProgressBarCallback
 
 pytestmark = pytest.mark.skipif(
     sys.version_info >= (3, 12), reason="Skip dask tests for Python version 3.12+"