Update dask tests to use a shared Ray instance fixture. One test requires a multi-cluster setup, so moved it to its own file.

Latest postmerge results:
```
[2025-05-21T05:31:19Z] //python/ray/util/dask/tests:test_dask_optimization                      PASSED in 7.9s
[2025-05-21T05:31:19Z] //python/ray/util/dask/tests:test_dask_optimization_client_mode          PASSED in 7.9s
[2025-05-21T05:31:19Z] //python/ray/util/dask/tests:test_dask_scheduler                         PASSED in 31.9s
[2025-05-21T05:31:19Z] //python/ray/util/dask/tests:test_dask_scheduler_client_mode             PASSED in 31.6s
[2025-05-21T05:31:19Z] //python/ray/util/dask/tests:test_dask_callback                           FLAKY, failed in 1 out of 2 in 60.7s
```

This premerge results:
```
[2025-05-21T14:38:33Z] //python/ray/util/dask/tests:test_dask_optimization                      PASSED in 11.7s
[2025-05-21T14:38:33Z] //python/ray/util/dask/tests:test_dask_optimization_client_mode          PASSED in 10.2s
[2025-05-21T14:38:42Z] //python/ray/util/dask/tests:test_dask_callback                          PASSED in 17.6s
[2025-05-21T14:38:42Z] //python/ray/util/dask/tests:test_dask_multi_node                        PASSED in 14.2s
[2025-05-21T14:38:42Z] //python/ray/util/dask/tests:test_dask_scheduler                         PASSED in 17.2s
[2025-05-21T14:38:42Z] //python/ray/util/dask/tests:test_dask_scheduler_client_mode             PASSED in 13.1s
```